pyproject.toml
[tool.poetry]
name = "django-drf-companyapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
celery = "^5.4.0"
black = "^24.4.0"
pillow = "^10.3.0"
gunicorn = "^22.0.0"
jinja2 = "^3.1.3"
drf-spectacular = "^0.27.2"
django-rest-knox = "^4.2.0"
django-rest-passwordreset = "^1.4.1"
django-filter = "22.1"
unittest-parametrize = "^1.4.0"
django-redis = "^5.4.0"
redis = "^5.0.4"
django-model-utils = "^4.5.1"
django-celery-beat = "^2.6.0"
workdays = "^1.4"
freezegun = "^1.5.1"
django-test-migrations = "1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
